temp commit

+++ backend/project/api/call/views.py
@@ -1,13 +1,16 @@
+from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.generics import ListAPIView, GenericAPIView
 
 from project.api.call.models import Call
 from project.api.call.serializer import CallSerializer
+from project.api.permissions import IsOwnerOrReadOnly
 
 
 class GetAllCalls(ListAPIView):
     """
-    GET: List all call
+    GET: Get the list of all the calls.
     """
     serializer_class = CallSerializer
     queryset = Call.objects.all().order_by('-created')
@@ -19,6 +22,9 @@ class GetAllCalls(ListAPIView):
 
 class CreateCall(GenericAPIView):
     serializer_class = CallSerializer
+    """
+    POST": Create a new call.
+    """
 
     def post(self, request):
         serializer = self.get_serializer(
@@ -32,32 +38,60 @@ class CreateCall(GenericAPIView):
 class GetUpdateDeleteCall(GenericAPIView):
     queryset = Call.objects.all()
     serializer_class = CallSerializer
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
+    """
+    GET: Get the details of a call by providing the id of the call.
+    """
     def get(self, request, **kwargs):
         call_id = self.kwargs.get('id')
-        call = Call.objects.get(id=call_id)
+        try:
+            call = Call.objects.get(id=call_id)
+            self.check_object_permissions(request, call)
+        except Call.DoesNotExist:
+            return Response(f'This call does not exist', status=status.HTTP_404_NOT_FOUND)
         serializer = CallSerializer(call)
         return Response(serializer.data)
 
+    """
+    PATCH: Update a call by id (only by call owner (organisation owner) or admin).
+    """
     def patch(self, request, **kwargs):
         call_id = self.kwargs.get('id')
-        call = Call.objects.get(id=call_id)
+        try:
+            call = Call.objects.get(id=call_id)
+            self.check_object_permissions(request, call)
+        except Call.DoesNotExist:
+            return Response(f'This call does not exist', status=status.HTTP_404_NOT_FOUND)
         serializer = CallSerializer(instance=call, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
 
+
+    """
+    DELETE: Delete a call by id (only by call owner (organisation owner) or admin).
+    """
     def delete(self, request, **kwargs):
         call_id = self.kwargs.get('id')
-        call = Call.objects.get(id=call_id)
-        call.delete()
-        return Response('call deleted')
+        try:
+            call = Call.objects.get(id=call_id)
+            self.check_object_permissions(request, call)
+        except Call.DoesNotExist:
+            return Response(f'This call does not exist', status=status.HTTP_404_NOT_FOUND)
+        if request.user.id == call.organisation.user.id or request.user.is_staff:
+            call.delete()
+            return Response('call deleted')
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class GetListCallOfVol(GenericAPIView):
     queryset = Call.objects.all()
     serializer_class = CallSerializer
 
+    """
+    GET: Get the list of all the calls a volunteer is participating.
+    """
     def get(self, request, **kwargs):
         vol_id = self.kwargs.get('vol_id')
         calls = Call.objects.filter(call_options__volunteers=vol_id)
